change Type Long to Integer

+++ study-room-reservation-system/src/main/java/com/test/studyroomreservationsystem/repository/ReservationUpdateDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 public class ReservationUpdateDto {
     private Enum<State> state;
-    private Long roomId;
-    private Long partitionId;
+    private Integer roomId;
+    private Integer partitionId;
     private Long reservationId;
     public ReservationUpdateDto() {
     }
@@ -15,7 +15,7 @@ public class ReservationUpdateDto {
         this.state = state;
     }
 
-    public ReservationUpdateDto(Long roomId, Long partitionId, Enum<State> state){
+    public ReservationUpdateDto(Integer roomId, Integer partitionId, Enum<State> state){
         this.roomId = roomId;
         this.partitionId = partitionId;
         this.state = state;
